Authorization with policies

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,12 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    
     /**
      * Display a listing of the resource.
      *
@@ -66,10 +72,13 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, 'Access denied');
         }
-
+        */
+       
+        $this->authorize("update", $question);
         return view("questions.edit", compact('question'));
     }
 
@@ -82,9 +91,13 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, 'Access denied');
         }
+        */
+
+        $this->authorize("update", $question);
 
         $question->update($request->only('title', 'body'));
 
@@ -99,10 +112,13 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        /*
         if (\Gate::denies('delete-question', $question)) {
             abort(403, 'Access denied');
         }
-
+        */
+       
+        $this->authorize("delete", $question);
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
